// models/profile.ts
export interface Profile {
  avatar: string
  name: string
  username: string
  github_username:string
  email: number
  skill: string
  bio: string
  role: string
  reputation: number
  badge_count: number
  comment_count: number
  tag_flowing_count: string
  post_count: string
  createdAt: string
  web_url:string,
  education:string,
  email_display:string
}